+++ server/src/EuroMotors.Domain/Orders/OrderItem.cs
@@ -14,7 +14,7 @@ public sealed class OrderItem : Entity
 
     public Guid ProductId { get; private set; }
 
-    public decimal Quantity { get; private set; }
+    public int Quantity { get; private set; }
 
     public decimal UnitPrice { get; private set; }
 
@@ -22,7 +22,7 @@ public sealed class OrderItem : Entity
 
     public Product Product { get; private set; } = null!;
 
-    public static OrderItem Create(Guid orderId, Guid productId, decimal quantity, decimal unitPrice)
+    public static OrderItem Create(Guid orderId, Guid productId, int quantity, decimal unitPrice)
     {
         return new OrderItem()
         {
